added ContosoExpenses ARM
This is a new project that deploys the entire Contoso Expenses App Service environment POC. This will be used in multiple POCs e.g. Application Insights. For customers that have already went through the App Service POC and now want to add additional features on top of that this is a quick to jumpstart their efforts.